rename from `fetch_example_datasets.py` to `fetch_examples.py`
